+++ yaqd_lightcon/_aiohttp.py
@@ -1,20 +1,7 @@
-from __future__ import annotations
-
 import aiohttp  # type: ignore
 import asyncio
 
-
-class TaskSet(set):
-    """container class for tasks to keep strong references"""
-
-    def add(self, task: asyncio.Task):
-        super().add(task)
-        task.add_done_callback(self.discard)
-
-    def add_coro(self, coro):
-        """create task and add to running loop"""
-        task = asyncio.get_running_loop().create_task(coro)
-        self.add(task)
+from ._taskset import TaskSet
 
 
 class Client:
